Refactor shared ecosystem types into src/pages/Resources/Ecosystem.ts
