+++ src/pages/Resources/Podcasts.data.ts
@@ -1,4 +1,4 @@
-import { Resource, ResourceCategory, ResourceType } from '../Resources';
+import { Resource, ResourceCategory, ResourceType } from './Ecosystem';
 
 const podcasts: Array<Resource> = [
   {
