+++ src/utils/rememberSearch.ts
@@ -1,6 +1,6 @@
 // Remember current search in page URL as ?search=...
 
-import { createEffect } from 'solid-js';
+import { Accessor, createEffect } from 'solid-js';
 import { useNavigate } from 'solid-app-router';
 
 export const rememberSearch = (search: Accessor<string>) => {
